+++ examples/standalone/src/apps/MyRunningMap/src/lib/geospatial/GeospatialManager.ts
@@ -1,0 +1,13 @@
+import type { Geolocation } from 'libmodulor';
+
+export interface GeospatialManagerDistance {
+    u: 'km';
+    v: number;
+}
+
+export interface GeospatialManager {
+    distanceBetween(
+        p1: Geolocation,
+        p2: Geolocation,
+    ): GeospatialManagerDistance;
+}