+++ examples/standalone/src/products/Main/manifest.ts
@@ -1,6 +1,6 @@
 import type { ProductManifest } from 'libmodulor';
 
 export const Manifest: ProductManifest = {
-    appReg: [{ name: 'Toolbox' }],
+    appReg: [{ name: 'MyRunningMap' }, { name: 'Toolbox' }],
     name: 'Main',
 };