+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Expose `UC_POLICY_FILE_NAME_EXT` and `UC_POLICY_FILE_NAME_SUFFIX` in the convention
 - Move `examples/embedded` to `examples/standalone`
 - Add the `GeocodeAddress` UCD to `Toolbox` (for docs)
+- Add the `MyRunningMap` app to `examples` (for docs)
 
 ## v0.19.0 (2025-06-20)